Updated imports, Added readme, Created file for tests
