+++ pyproject.toml
@@ -9,6 +9,7 @@ readme = "README.md"
 python = "^3.8"
 selenium = "^3.141"
 pytest = "^6.2"
+pytest-html = "^3.2.0"
 
 
 [build-system]
